+++ src/main/java/likelion/MZConnent/domain/culture/CultureCategory.java
@@ -1,8 +1,12 @@
 package likelion.MZConnent.domain.culture;
 
 import jakarta.persistence.*;
+import likelion.MZConnent.domain.member.SelfIntroduction;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Entity
 @Getter
@@ -15,4 +19,7 @@ public class CultureCategory {
 
     @Column(length = 255, nullable = false)
     private String name;
+
+    @OneToMany(mappedBy = "cultureCategory", cascade = CascadeType.ALL)
+    private List<SelfIntroduction> selfIntroductions = new ArrayList<>();
 }
